refactor: extract useRowCollection + RowItem from the Repeater renderer
